+++ libfive/src/render/brep/simplex/simplex_mesher.cpp
@@ -489,7 +489,7 @@ uint64_t SimplexMesher::searchEdge(Eigen::Vector3d inside,
     // TODO: we should weight the exact position based on values
     Eigen::Vector3d vert = (inside + outside) / 2;
 
-    return m.pushVertex(vert.template cast<float>());
+    return m.pushVertex(vert);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
